+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add the `IMGPROXY_MAX_REDIRECTS` config.
 - (pro) Add the `IMGPROXY_SERVER_NAME` config.
 - (pro) Add the `IMGPROXY_HEALTH_CHECK_MESSAGE` config.
+- Add the `IMGPROXY_HEALTH_CHECK_PATH` config.
 
 ## [3.2.2] - 2022-02-08
 ### Fix